Final Project Game
I’ve been having a few issues with getting collision to work between
the four colored balls and the four colored chambers so that when a
proper-colored ball collides with its respective-colored chamber, the
ball disappears; when a certain-colored ball collides with a
different-colored chamber, it results in a game over. Also wanted to
add the player’s ability to close off one chamber at a time by using
key presses.